requirements.txt
python-gitlab>=3.0.0,<5.0.0
python-dotenv>=1.0.0
PyQt6>=6.0.0